Decouple gravity from the cluster pgen into src/gravity
ClusterGravity (pgen/cluster/cluster_gravity.hpp) implemented an NFW halo +
BCG (Hernquist) + point-mass SMBH radial profile with nothing actually
cluster-specific about the physics. The only cluster-specific parts were a
hardcoded "problem/cluster/gravity" input block and a `friend class
SNIAFeedback` letting SNIAFeedback reach into private members to disable
NFW/SMBH for its BCG-only density estimate. This is groundwork for reusing
the same gravity machinery in a future non-cluster (spiral-galaxy) problem
generator.

- New src/gravity/spherical_gravity.hpp: the NFW+BCG+SMBH implementation
  moved verbatim into a generic gravity::SphericalGravity, with two changes:
  the input block name is now a required constructor argument (block_prefix)
  instead of hardcoded, and the friend-class hack is replaced by public
  DisableNFW()/DisableSMBH()/WhichBCG() methods. The module now has no
  knowledge of SNIAFeedback or "cluster" at all.

- pgen/cluster/cluster_gravity.hpp shrinks to a ~40-line shim:
  cluster::ClusterGravity : public gravity::SphericalGravity, fixed to
  "problem/cluster/gravity". Same namespace, type name, both constructor
  signatures, and "cluster_gravity" Param registration name as before, so
  every other call site (cluster.cpp, hydrostatic_equilibrium_sphere.*,
  snia_feedback.*) and every existing <problem/cluster/gravity> input block
  keeps working unchanged.

- snia_feedback.cpp updated to call the new public Disable*()/WhichBCG()
  methods instead of the removed friend-based private-field access.

Verified: full rebuild compiles clean; cluster_hse, cluster_tabular_cooling,
cluster_magnetic_tower, cluster_hydro_agn_feedback, and cluster_agn_triggering
regression tests all pass unchanged.
